feat: remove check command and supporting code
